+++ tslib/src/functions/check-allocator/PupilPinGenerationService.ts
@@ -1,0 +1,9 @@
+export interface IPupilPinGenerationService {
+  generate (): number
+}
+
+export class PupilPinGenerationService implements IPupilPinGenerationService {
+  generate (): number {
+    throw new Error('not implemented')
+  }
+}
